// src/app/core/models/category.model.ts
export interface Category {
    id:               string;
    name:             string;
    description:      string;
    type:             string;
    code:             string;
    state:            null;
    isActive:         boolean;
    createdBy:        null;
    createdDate:      null;
    lastModifiedBy:   null;
    lastModifiedDate: null;
    companyId:        string;
}